flight detail huge update

+++ src/main/java/com/dmytrobozhor/airlinereservationservice/service/AbstractFlightDetailService.java
@@ -2,6 +2,6 @@ package com.dmytrobozhor.airlinereservationservice.service;
 
 import com.dmytrobozhor.airlinereservationservice.domain.FlightDetail;
 
-public interface AbstractFlightDetailService extends AbstractCrudService<FlightDetail, Integer> {
+public interface AbstractFlightDetailService extends AbstractCrudService<FlightDetail, Long> {
 
 }
